agent_aws: simplify period validation
The parameter "Period" is the granularity of the data returned by the
AWS API. This parameter has a few constraints: in most cases it should
be a multiple of 60 (see
https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_MetricStat.html).

The granularity was previously set using each endpoint's caching
interval. It is a different property though, and is now set
independently. It is therefore not necessary to ensure that the caching
interval meets the constraints of the "Period" parameter.

CMK-5838

Change-Id: I2ae4551491a656e6f12541e09b2fadb7137ccf33