Fix host attribute update caching
After completing the function, the hash was not stored correctly, making
the request execute the whole function on every request which takes
several seconds with larger configurations.

The config hash is now stored correctly again. Moved the config hash to
a dedicated variable. This global state is still not ideal, but at least
our tooling can protect us better from typos compared to the previous
approach.

Change-Id: Ie250a5f99129f037d808fce2fb642ffe104f4a66