11333 FIX aws_agent: enable cache to reduce live data requests
The built-in hash function is used to create a hash from the agent's
input arguments, which is saved to a config file. This is to understand
whether any of the agent's config has changed, and live data needs to be
retrieved. This function produces a different hash for the same input
arguments across different python processes, which the agent interprets
as a changed config. The hashing mechanism has been replaced to ensure
that the same input args always lead to the same hash.

CMK-5838

Change-Id: I69df9ce1f029b5ea4574f05a1d7d4d6becf0c29f