Fix bug displaying single metric dashlet single value
Because transformed data is sorted it happens every second minute that the
current value has a timestamp lower that the latest value returned by the
rrd. Out of that reason values would be displayed every 2 minutes.

find picks the first element that has the entry formatted_value, that is
the entry of the current value. Search is linear and only over the 2 first
elements as the array is sorted.

Change-Id: I5fa6f30650f586cbe3dee7720a4b6e3cd82d7699