Figure dashlets don't need to drop title as default
Showing title is a general default of dashlets and now that the rendering
is unified figure dashlets should also render the title by default.

The default of a title is the dashlet title, plain, which gets overwritten
by custom title when set. On Figure dashlets, because the dashlet title is
not sent over ajax, as it is not part of the spec, it returns to the
previous behavior of not having a dashlet title shown until custom title is
entered.

Change-Id: I9428fc8ba21011695376314669945b19dbdbe41f