Revisit metric reverse translation update strategy
With cmk 2.0, it is now possible to migrate metric names and query
simultaneously old and new metric name. For performance graphs it takes
little overhead, because only 2 metric names are queried. For the Metric
history painters is a bigger performance gain if less metric names have to
be queries, thus the migration took place. To reduce the queried names.

Goal is/was to migrate then all checks to use their canonical metric name
by cmk 2.0.0, so that by cmk 2.1.0, we don't need to reverse translate and
have a fast performing metric history painter.

This change represents that goal, by checking that from version of
deprecation plus one stable version the old metric name can still be
queried. This gives users enough time(1 year) to collect data under the new
metric name and not feel a disruption on the feature.

The previous test, assumed we write the version of deprecation, when the
metric shall be removed, yet we write the version when the metric got
unified and no longer requires a reverse translation.

Change-Id: Ia9ed9a9293a2bb17df95801a2fd98d4a772f57c9