Improve consistency btw `do_check` and `check_discovery`
Both functions may be used interchangeably in parts of our code, such as
in the DiscoCheckExecutor so that the added consistency helps to avoid
bugs.  Moreover, the last two arguments to `do_check()` are optional
sequences so that swapping them may remain undetected by the type checker.
Making them keyword only fixes this issue.

CMK-5815

Change-Id: I65507f39d42d0aab30c1bfff32c65cffbc465705