Extend test for reverse translation
Check-plugins should be migrated to deliver metrics under their canonical
name. We mark on the translation on which version the migration happened
and keep the reverse translation valid for an entire stable version
release.

Thus test needs to be able to pin the version to verify the translation is
taking place or being already dropped.

It would be good to manage to migrate all memory metrics before 2.0.0
release. Here is a test over mem_linux still having unmigrated
metrics(shmem).

Change-Id: Id8596db99e65583b99a59035a9821f2aaf0919be