11505 FIX Active checks: Failed check caused by missing arguments
In Checkmk 2.0.0i1 several active checks failed with different error messages,
because the command lines that were built to execute the active checks were
missing required arguments.

For example the SMTP checks failed when a custom network port was configured
with a message like: "check_smtp: Port must be a positive integer". Since this
was an error in a generic place, this affected multiple active checks.

Change-Id: I51c78a0e8914846d6bdb2bd7d42b6cdc206f9823