Rewrite WATO page handler to PRG pattern
Instead of "in page redirect" to another mode, we now use the
POST/redirect/GET pattern. This prevents problems with page
reloads.

This change will break several pages in WATO. We'll fix this
in the next commits where we simplify the ActionResult type
by extracting the different features from it until we end up
with an Optional[FinalizeRequest] type.

Change-Id: Ifda23e09e382604ff2c6ee95d7090e966a496632