Check-API: decouple API from implementation: SNMPDetectSpecification
Split up the classes of the SNMP detect specification in two:

The class defined in the API only has the job of hiding the implementation
of the detect specification: It is the argument type of the "detect"
keyword of the SNMP section registration function, and the return type
of utility functions such as any_of, all_of, startswith and others.
As users are not expected to create detect specifications directly,
their exact nature is an implementation detail.

The other class is used in the backend code, and (currently) provides
methods for (de)serialization.

They are currently linked by the common base type (to avoid casting),
but as long as we can create the backend class from the API object,
we are free to change things.

CMK-6136

Change-Id: I98e36ec8876a6922991118c70b38abfe8a669e50