11509 FIX Fix password change confirm dialog
When trying to change the login password using the GUI, the error message
"Invalid credentials" was shown and the user was logged out. After that, the
user could login with the new password.

This has now been fixed: After changing the password, a confirmation is shown
and the user does not need to login again.

Change-Id: I6e5c6bb73d7faa2ac1e22c65dd5d82cf60e4af73