SNMPTree: move API code to API
We split up the SNMPTree object into to separate Objects, one that
is exposed by the API (featutring validation), and one that is
used by the backend (featuring de-/serialization).

That way we can move each functionality where it belongs, and
change the backend code without effecting the API.

Additionally some missed renamings of SNMPDetectSpec to
SNMPDetectSpecification are fixed in the tests.

CMK-6136

Change-Id: Ib7546e391d342eee44a627d91b1c126e57034db7