Enforce Host > Service description hierarchy in Context / Search filters
Context / Search filters are a generic and present filters ordered if they
are single info or multi info. The only order specified comes by the list
order given by the info handler and only affects the single info part.

Checkmk is based on the hierarchical ordering of host>service>rest, thus we
enforce that on the display of those filters.

CMK-6011

Change-Id: I20e38b7723f8cedbfd46ea9252053ca4d96b8ff8