The status border for the single metric dashlets is toggle-able
Single metric, gauge and Bar, dashlets use single metric datagenerator. On
their options they can toggle to display a status border, not on the
general dashlet settings. Could be moved later, as the config is flat
between settings and properties.

Bar dashlets can be multi host. It remains to be specified what the border
shall mean in that case. and if it shall be solved on front end or backend.

CMK-5871

Change-Id: I3f2ff010bcf3831f01baf7cad3ca9a8e14c7dbf8