Introduce new mechanism for confirm dialogs
Form resubmission issues in WATO have been fixed for regular forms
recently. This problem is currently unsolved for forms involving confirm
dialogs.

This commit introduces a solution for that issue by replacing the
existing confirm dialog logic with a Javascript based confirm dialog.

This has the advantage that the confirm message is raised before the
form is submitted to the server. It really reduces the complexity of the
confirm dialog logic and will make it possible to simplify our Python
code.

Some drawbacks have already identified: Previously the confirm dialogs
were rendered by the Python code. This made it possible to

a) perform some logic on the Python side to get more information for the
   confirm message.
b) perform some of our server side validations before displaying the
   confirm dialog.

After a quick look at some call sites it looks like we can live with it.
Let's go with this solution and solve possible problems if we stumble
upon it.

CMK-6214

Change-Id: Idb99369a1cdb3f857e12ceba5e7285a7705620f9