Add confirmed form submit helper for page menu submit links
In case we have forms with multiple different actions, where some of
the actions need a confirmation and some not, we need to have an
individual handling for each button.

This helper function supports this case by providing an individuall helper
function for all page menu entries of this kind.

CMK-6214

Change-Id: I7ea68ac51700cd4e8e2ff71adbbe734a38b296c3