deb: Move compilation from binary to build phase
The binary phase is executed in context of "fakeroot" (see "man
dpkg-buildpackage") for further information. In the past we used this
phase for building our packages, but we experienced some hardly
reproducible random IO errors when executing our build in this phase.
There are indications that these IO problems have to do with the
fakeroot.

We also had some noisy LD_PRELOAD related warning messages in the build
log which were also caused by building in fakeroot context.

To solve these issues we now build our stuff in the build-arch phase
which is executed in the current users context.

Change-Id: Icd79e802c610f0c4e81677c93c3f68487b537a99