11465 FIX Fixed race condition which may cause crash during start of Microcore
There was a race condition in the Microcore code that may cause crash at the start.
The log usually contains at the very beginning of the Microcore start something like
"Terminated due to unexpected signal 11. Restarting." and "stack trace with 17 frames
follows".

With this release the error has been eliminated.

Change-Id: I87838a2c3f4cd450524a969febf51525be955357