11622 FIX esx_vsphere_counters
This werk fixes a bug in the legacy check esx_vsphere_counters, which
prevented this check from being discovered and crashed already
discovered services.

------------------------------------------------------------------------

Internal note:

Most legacy checks import the needed legacy-includes via *-imports. This
is the correct way in this case, since it ensures that all needed
variables are present when resolving legacy discovery results such as
[("SUMMARY", "diskstat_default_levels")].

However, for some legacy checks, the *-imports were replaced with more
specific imports, which can cause problems due to the reason explained
above. This change re-introduces the *-imports for these cases.

Change-Id: Ia6fc999f8f6c8cda1e9727db73db35d53873c696