Move html.confirm to dedicated module
This function should now only be used in very special
situations where we want to show some information from
the Python code and give the user some preview before
confirming an action.

Move it out of the way, rename and document it to make
this clear. Also simplify the method by removing unused
code paths.

Change-Id: Ie384707ef576dc4efb07cd88691305538d4e8432