10609 Simplified display of pending changes in setup
The header information on setup pages (in the top right corner of the page) was simplified.
It now shows only the number of pending changes and links to the "Activate pending changes" page.
The information about (selected) rows can now be found on the top right of the respective table.
(CMK-5583)

Change-Id: I1ffe87e6fb115a3dcba9988b0d186ec0e1091cc7