Group {de,}serialization of GlobalConfig / Fetcher
This way, we let the type system make sure that writing and reading do
not get out of sync.  This further helps understanding the code and
lets us modify/extend the configuration as needs be.

We further SRP `load_global_config` and EAFP it as well as it is the
recommended way to handle files in Python.

Lastly, we remove an uninteresting test.

CMK-6181

Change-Id: I06057a1595d4bc486c49e55d848fbf27db9e026f