11625 FIX Splunk special agent: fix section <tt>splunk_jobs</tt>
This werk fixes a bug in the section <tt>splunk_jobs</tt> of the
Splunk special agent. This bug caused the agent to crash while
producing the output for this section. As a result, the section
<tt>splunk_jobs</tt> and any subsequent sections were not fetched,
which made the corresponding services go stale (e.g. "Splunk
Alerts" and "Splunk Health").

Change-Id: I0e20584677a012994e144a0f0b092b6fabf4c4a2