Clean up "--checks" option
The "--checks" option of the cmk command is meant to restrict execution
of check plugins to the listed ones.
In the past, this also had the undocumented effect of changing the 'scan'
behaviour of SNMP plugins. This effect has disappeared during fetcher
development, and was not always helpful anyway. It will be implemented
as a separate option.

The "--checks" option now strictly restricts execution of discovery
and checking to the selected plugins.

The same feature is prepared for in the inventory module.

Change-Id: Iba4f35bb84e0557009560657698fafba885283e6