Add object diff capability to audit log
The add_change and log_audit functions now accept two additional
arguments old_object and new_object. These are the objects we use
for building a description of the changes.

These changes are written to the audit log in text form. This has to
done in this situation, because we don't want to persist the whole
object structures (to keep the size of the audit log small).

Recording and displaying object differences is an optional feature
of the audit log for now.

CMK-6187

Change-Id: I0be8afd866a23e0528cf7bfd29de7099916d14ab