Simplify audit log text transport
We can now use complex data structures and don't have to wrap the whole
audit log entry into a single line.

This makes it possible to move the escaping of the message to the
rendering as we usually do. We just need to serialize deserialze the
HTML object information to make this possible.

Change-Id: I8a6b46b214588e6b33c87c07ca3434fecfa22010