Cleanup typing of automation discovery (aka inventory) call
We now have the type AutomationDiscoveryResult and helper types that
clarify the typing of the involved data structures in the Base and GUI
code base.

This is a preparation for extending the exchanged data with a text that
describes the changes to the services, which will then be written to the
WATO audit log.

CMK-6332

Change-Id: I604d93eaf173f3e17b3a656e9a40dc6ec5a739de