Extract object diff calculation from log functions
Turns out it is better to keep the diff text calculation separate from
the change log entry logging. The separation allows us to get the diff
text from another process without the need to transport the whole old
and new data structures.

This is useful when automations, which have the modified data structure,
are executed on remote sites and have to respond with the diff. This is
the case with the "inventory" automation.

CMK-6332

Change-Id: I9d2dd4ed32b7572bb1b5609863c68b822a714602