Performance graph title format transformation
When graphs are rendered async the graph_render_options arrive from a JSON
object and the Tuple type is not present anymore. At the moment problem
appeared in graph collections which have a fixed not subject to
customization title_format, updating the specification already fixed the
problem.

This is just an extra cautionary test on the transformation.

CMK-6339

Change-Id: Ib172d80b657520cc6d788de42018d869dc3262cd