11409 Add service label generation to network interface discovery rule
As an addition to the reworked "Network interface and switch port discovery"
ruleset, as described in Werk #11361, it is now possible to automatically
generate service labels via discovery.

To achieve this, you can specify a set of service labels per created "Network
interface and switch port discovery" rule, that will be applied to the matching
single interfaces and interface groups on discovery.

In order to apply the configured labels, you have to (re)discover the services
that correspond to the matching interfaces and interface groups

CMK-4632

Change-Id: I819073a12c8eff466783c54141e5bd375ceb5ebc