Refactor audit log and site changes object references
* Both supported only host and folder references
* The audit log previously used the format "folder_path:host_name" as
  reference and named it "linkinfo".
* The site changes were persisting the reference as two element tuple
  like this: ("CREHost", "hostname"), ("CREFolder", "/folder/path").

This change now consolidates both to a common data structure. During
runtime it is represented by the ObjectRef object. For persistence,
the ObjectRef is serialized and deserialized in a standardized way.

Wit the common data structure, this change also unifies the rendering of
the object titles and URLs. The audit log now has a dedicated column for
the object type.

Following changes will add object references for additional object
types.

CMK-6332

Change-Id: Ie626bc2c286b83d640b8a09114da24b8741cdf01