10610 Simplified design of page reload
The page reload information in the upper right corner of view headers was simplified.
The number of seconds of one reload interval is shown in the center of a circle.
This circle grows clockwise and thereby visually displays the remaining time until the next reload.
Clicking on the number/circle triggers an instant reload.

Change-Id: I4124ef4d489d8632bf2f3954f11f931bbe95bdd6