Idle CPU: Filter negative values
In some cases with short measurements, it is possible that the errors in
the process measurements accumulate enough to be slightly higher than
the elapsed time.  In this case, `Snapshot.idle` may return a negative
value in the order of a few ms.

This patch filters negative idle values out.

CMK-6259

Change-Id: Ie09e671ad73f7673ebc708587a011861385241d7