Make rules identifiable by unique IDs
Since the early days of Checkmk the rules were always referenced by
their position in the rule set list plus the related folder.

We have a problem with rules not being clearly identifiable:

* Audit log only refers to *current* list position
* Possible race conditions in "edit ruleset" actions
* API calls can not be made directly on individual rules

The new ID will be created once for all existing rules during update to
Checkmk 2.0 and persisted for the whole lifetime of the rule.

CMK-6369

Change-Id: I7188e8deaf5c2b86a85d0a64d6f88b202e57c1ac