fetcher protocol: Add CMCMessage class
The class is consistent with other classes in the module where a
`*Message` class contains the corresponding header and payload.

This class also let's us write a parser for the CMC messages.

This way, we can finally modify the tests so that they do not compare
bytes but human-readable messages.

Change-Id: I2e501bededf128a0fc32f738b0d58e4d9b004589