[CMK-6318] - add host name to the message to make a error message clear
- unskip and fix unit tests.

Tests should be improved(better structure and better fixtures) still at the moment it is not urgent and may be performed later(in background).

Another point of possible improvement is changing interface to enable_timeout to more consistent(with Command, for example, to have readable error as one unified structure). Again, this is not urgent and a bit out of ticket scope.

Change-Id: Ib3c98624ab1f2f10305563f151c99d1fef3eb612