Single metric dashlet areaplot with solid translucent fill instead of gradient
In UX dashboard there are only two colors. Light blue line and gray. Yet
our metric system has a color for each metric. That color is kept as
default and the user is capable of choosing any other color. This is
simpler and more extensible that setting a dropdownchoice for 2 options.

Default is to use metric system color, if that fails take the UX light blue

Let next UX workshop decide about this or really 2 color.

CMK-5649

Change-Id: I0d7528fa8939b93aac3795538190f66e513642b9