PersistedSection typing: Remove unnecessary types
This patch improves the type hierarchy to PersistedSection by removing:

 - {Agent,SNMP,T}PersistedSection
 - {Agent,SNMP,T}PersistedSections

The hierarchy was duplicated for the Agent and for SNMP.  The
duplication then obscured the actual differences and parametrization of
the types to the users, eventually leading to more duplication.

The intent of the code is also clearer now that the user has to
explicitly parametrize the PersistedSection type.

CMK-6362

Change-Id: Ia6d00b94b43b42fe217b6597e40a7d63c57fbb78