11675 Developer option: --detect-plugins
This werk adds the new commandline option "<tt>--detect-plugins</tt>" to the <tt>cmk -I my_host</tt>,
<tt>cmk -i my_host</tt> and <tt>cmk my_host</tt> commands.

It is compatible to the previous "--checks" option, which will still work.

Setting this option to a comma separated list of plugin names will have the effect that the specified
command (that is, checking, inventory or discovery) will only be done for the provided plugins,
in the same way "--plugins" would.

In addition, checkmk makes a besst effort guess which sections should be fetched, that is:
It sets the "--detect-sections" option to all sections that may be needed for the provided plugins.

Change-Id: Ibe017a6be2d8deab9b0689cc967876ae77459b20