11380 FIX mk_oracle: Fixed reference to all found SIDs in custom sqls
In the help text is the assignement of <i>SQLS_SIDS="\$SIDS"</i> an example
to assign a custom sql to all found instances. This has not been working,
since the config loaded before the script looked for running instances. The
example configuration of the help text will work with this fix.

Change-Id: Icfe3422fcc1b38b45776ad73f4f62cd9176479d0