11610 FIX Bulk import: Improve validation of imported attributes
When setting e.g. tag attributes during the bulk import, the values from the
CSV file were not validated using the normal validation logic of the host
attribute. This made it possible to import attributed with invalid values,
leading to errors in different situations.

The import now applies the same validation logic as the host edit forms and
prevent importing data which does not pass this validation.

CMK-5473

Change-Id: If49c500fbfa53949c29e29d86fa6568438bbb6c3