11676 FIX esx_vsphere_vm_snapshots: New display defaults for summary service
The display of the 'ESX Snapshots' and "ESX Snapshots Summary" was half-baked.

The "ESX Snapshots" service is now unconditionally displayed on the VM related to the
snapshots. The "ESX Snapshots Summary" service is created for the vCenter by default.
Users who do not have a vCenter included in their monitoring (or at all) can choose to
display the summary service on their respective ESX hosts by activating the check box
"Display snapshot summary on ESX hosts" in the agents rule ("VMWare ESX via vSphere").

Change-Id: I0882ddceff899825c15e6515e0a21935036ea317