Audit log labels: Cleanup rule references
Previously the ruleset name was merged into the ident, because there
was no better way to transport the information, which is currently
only needed for the link to the rule page.

This commit introduces an optional labels field, which can be used
to transport object specific information values for audit log entries.

Add some tests for the object reference serialization.

CMK-6444

Change-Id: Ieda3114bd95a8b929c1b7e32c3fa55151c5872a2