11730 FIX oracle_jobs: correct usage of user-defined parameters for missing jobs
The check plugin <tt>oracle_jobs</tt> now responds as expected to user-defined
settings for missing jobs (ruleset "Oracle Scheduler Job", option "Status of
service in case of missing job"). Previously, this settings was not used.
Furthermore, the default service state for missing jobs is now CRIT (which
matches the default setting in the ruleset), before, it was UNKN. Note that
for this new default setting to have an effect, a re-discovery is necessary.

This merges PR #279, https://github.com/tribe29/checkmk/pull/279.

Change-Id: I05a9e4abb6f50166d580871fdacb2eac1d9e9ead