agent parser: Remove one unnecessary tmp variable
We remove the tmp: `section_content` and move the logic
to `section_header`.  Indeed, the parser "knows" it is
parsing a host section because it could obtain a
host section header.  That there may be a section content
is but a side effect.

The improved logic advantageously replaces the logic of
the state machine from crash to noop.

CMK-5004

Change-Id: Ifd0d18b473820b97c8aa02d4406572216159b39b