11679 FIX Wrong results of Check_MK Discovery service
In Checkmk versions 2.0.0i1 and 2.0.0b1 the core config was missing some discovery rules.
This resulted in some wrong rusults of the "Check_MK Discovery" service (in particular one 'df'
service was wronlgy reported as vanished).

Internal note:
The missing copy.copy had the effect that the values where updated not
only in the config variable, but simultaneously in the check_default_variables.
As a result the config variables compared equal to the default variables
in `PackedConfigGenerator.generate`, and were not added to the config.

Change-Id: Ie0e1c1da736c3040b63e1e4ee568ef4e0f599dbb