Deal with Grafana Datasource empty site filter on combined graphs
Grafana default dropdown for site filter, when no sites are selected, is an
emptry string. Livestatus was filtering against an emptry string site and
thus no results could be found, because context['site'] is site enforced.
Grafana connector is correspondingly updated to use siteopt

Graph template ids are strings, the not found exception needs to match that
option.

CMK-6127

Change-Id: Ifdf0e4456bb9b7acdedbe64477d10231227ac19e