11732 FIX if-checks: update default check parameters
This werk updates the default check parameters of the interface checks.
This has no impact on the produced check results, but it fixes a bug in the
parameters view of the corresponding services ("Parameters for this
service"). Due to the default parameters being in an outdated format, this
view sometimes showed incorrect check parameters.

FEED-5356
CMK-6472

Change-Id: Iddbf03d41eaf73385edf5a10ea54e52bb468764b