11711 FIX mk_logwatch for Windows
If the mk_logwatch agent plugin was used no Agent Output was created and
the HW-/SW Inventory would fail aswell. Therefore the agent plugin could
not be used for Windows.

The reasons for that were:

* In Python 3 sys.stdout returns \r\n as linebreak for Windows.
Checkmk can't handle this therefore we removed it by replacing
the default sys.stdout.
* The win32api module could not be imported because pywin32 was not installed.
Now pywin32 is installed by default.

Change-Id: I670eefb86b54f89b0464617f1e24fcfc00fdd396