11711 FIX mk_logwatch for Windows
If the mk_logwatch agent plugin was used no Agent Output was created and
the HW-/SW Inventory would fail aswell. Therefore the agent plugin could
not be used for Windows.

The reasons for that were:

* In Python 3 sys.stdout returns \r\n as linebreak for Windows.
Checkmk can't handle this therefore we removed it by replacing
the default sys.stdout.
* The win32api module could not be imported because pywin32 was not installed.
Now pywin32 is installed by default.

Changes related to this werk:
* Fix \r in agent output    : 33ec94a8cf82a7b75e1350e1a5f83223c6a3bfe6
* Increase BUILD_NUM        : 634f749eacd06df08c3548a50589c1e8c9f527ea
* pypiwin32 -> pywin32      : b19cc9fd2d5c605500b9bd017abea0edf4a461ce
* pypiwin32 to Pipfile      : c50bbc7dc489331b9cbef35f7a7768f96c00bed4

Change-Id: I670eefb86b54f89b0464617f1e24fcfc00fdd396