11747 SEC Fix stored XSS triggered by received syslog messages
You are only affected by this issue in case you use the Event Console.

An attacker could send messages to the Event Console, e.g. via syslog,
containing arbitrary HTML code. This was executed in the browser context of any
user viewing the event in the Checkmk user interface.

The information is now properly escaped in a generic way to prevent these
issues.

Change-Id: I5d4f3594e69de7980aa474b8e3b9aa94d7342bb2