Try to fix flaky composition tests
The test was configuring a new site. This setting is automatically
applied to the Livestatus proxy daemon by sending SIGUSR1 to the
process. However, Checkmk is not waiting for the process to apply the
configuration while the test was happily continuing with it's logic.
Perfect race condition.

This commit extracts some existing wait logic from liveproxyd related
integration tests and also uses it during the composition test setup.

CMK-6614

Change-Id: I9824c6060ab1f9c4bccc5d61dd25d425f8fc37a0