11712 FIX mk_jolokia: Fix plugin execution on Windows
If the mk_jolokia agent plugin was used the HW-/SW Inventory would
fail. Therefore the agent plugin could not be used for Windows.

The reason for that was that in Python 3 sys.stdout returns "\r\n"
as linebreak for Windows. Checkmk can't handle this therefore we
removed it by replacing the default sys.stdout.

Changes related to this werk:
 * 33ec94a8cf82a7b75e1350e1a5f83223c6a3bfe6

Change-Id: I64b3bda8042c192995060cccf6273edd239709e1