Fall back to C++17 for now, clang doesn't like GCC-10's headers.
The current state of C++20 support regarding compilers and headers is a bit
confusing: GCC 10 and its accompanying headers partially implement C++20,
and Clang++ 10...12 has partial C++20 support, too. But there is some stuff
in GCC's headers that confuses Clang++, mostly regarding comparison operators.

Because of this, we can't use Clang-based tools in C++20 mode, so we fall
back to C++17.  This isn't that bad, lots of interesting C++20 stuff is
still missing, anyway.

Change-Id: I8ab8f86118ab5963cf91dd479b9d34e478f8ee37