11735 FIX <tt>arc_raid_status</tt>: Do not go CRIT when checking raid consistency
The check plugin <tt>arc_raid_status</tt>, which monitors Areca Raid Controllers,
reported the monitoring state CRIT when the raid was in the state "Checking".
However, this state means that a consistency check is currently performed on the
raid and does not indicate any issues. Therefore, with this werk, the monitoring
state of raids with the status "Checking" will be OK.

This merges PR #283, https://github.com/tribe29/checkmk/pull/283.

Change-Id: I9b7074b127cf1414b090d601a53a19e177abaf76