11685 FIX check_mk_agent.linux: moving sections exclude rule option to linux agent section
In the previous version, the "Exclude specific agent sections" rule was listed in the Checkmk
Agent section under the Monitoring Agents WATO page. This led to confusion as the option is only
available for users who have access to the Agent Bakery functionality. This werk moves the option
to the Linux Agent section to resolve this.

Change-Id: I94e6ea3aa9b456ce6bebc54a9943c35ce73d5ea9