Cleanup URL prefix from cmk-update-config errors
The cmk-update-config script uses a RequestContext from the GUI
code which prefixes the ERROR (or above) log entries with the
URL http://localhost/. This may be useful in the context of the
regular GUI, but not during the update script.

Quick fix this by introducing a flag. Obviously this should be
solved in a cleaner way, extracting all the log patching stuff
from the RequestContext.

Change-Id: I23a078a982639ae046e71a5c2d5e9b8685842576