11772 Bring back Agent/SNMP info in `cmk --list-checks`
The in Checkmk versions 2.0.0i1 and 2.0.0b1 the TCP/SNMP information
had been dropped from the output of <tt>cmk --list-checks</tt>.

This was due to the fact that check plugins are now neither purely
one or the other.
We now write "tcp" if the check plugin requires an agent data source
(e.g. the regular Checkmk linux agent), "snmp" if it reguires a SNMP
data source, and "agent/snmp" if it can be discovered with either of
the data source types.

CMK-6693

Change-Id: If368d92370e36061672a23aa5d9390bfe188da90