Revert "Revert "Pass `snmp_fetch_interval` to fetcher""
Currently, we use the second field from the value of the PersistedSections
to decide whether the cache should be invalidated or not.

This field contains the unix time computed from the first field (the
unix time when the data was cached) and the fetch interval at the time
the value was cached.  However, the computation is not updated when the
user modifies the fetch interval.  There is therefore no guarantee that
the value from the DB is correct at any time after caching.

Instead of using the value in the DB, we must always use the currently
configured interval and we therefore:

1. pass the configured values to the fecther (this patch) and
2. compute the invalidation date from the first field and the
current interval.

Actually, the second field should *never* be used.

CMK-6516

This reverts commit e0940cecea445c2478bb1af6fcaa562779265099.

Change-Id: I91727e42e38880debb0e00f3368ef24296e26ecf