11736 FIX Rename hosts: folders names for changed rules in audit log
When renaming hosts, Checkmk adjusts the conditions of rules which
include the renamed host. Such changes are recorded in the audit log,
including the name of the folders which contain the adjusted rules.
Up to now, these folder names were not displayed correctly. The audit
log showed something like
"Renamed host in 3 rulesets of folder <bound method CREFolder.title of
Folder('', u'Main directory')>".
With this werk, this becomes
"Renamed host in 3 rulesets of folder Main directory".

Note that this bug only affects version 1.6 and is already fixed in
version 2.0.

This merges PR #286, https://github.com/tribe29/checkmk/pull/286.

Change-Id: I9c3227640561578657b805b0a7163406dd0abaf2