11762 FIX Site attribute of hosts / folders is now limited to distributed setup sites
In the host and folder attribute configuration, previous Checkmk versions
allowed the user to select all configured remote sites. This included status
only sites and remote sites Checkmk is replicating the configuration to.

The setup is not able to replicate configurations to status only sites. For
this reason it does not make sense to configure status only sites on hosts or
folders. This change now excludes the status only sites from these attributes.

Change-Id: I7154b0d1a45b8a8fbaa0dbf7cffd0b5b01122039